udemy/python/requirements.txt
pylint
requests
lxml
bs4
pillow